GPB-47 - React code cleanning
