--- gpb-front/src/components/GameImage.js
@@ -1,25 +1,0 @@
-import * as React from 'react'
-import defaultImage from '../img/defaultGameImage.jpg';
-import Message from './Message';
-
-
-export function GameImage(props) {
-  let image
-  try {
-    image = require(`../img/games/${props.gameName}.jpg`)
-  } catch {
-    image = defaultImage
-  }
-
-  return (<img class={props.className} src={image} on ></img>);
-}
-
-export function GameAvailability(props) {
-  return (
-    <div class={props.available ? "App-game-content-list-game-info-available"
-      : "App-game-content-list-game-info-available not-available"}>
-      {props.available ? <Message string={'app.game.is.available'} />
-        : <Message string={'app.game.not.available'} />}
-    </div>
-  )
-}
