+++ gpb-front/src/components/common/Footer.js
@@ -1,6 +1,6 @@
 import * as React from 'react';
 import Localization from './Localization';
-import Message from './Message';
+import Message from '../../util/message';
 
 export default class Footer extends React.Component {
