+++ gpb-front/src/components/common/Header.js
@@ -1,8 +1,8 @@
 import * as React from 'react';
 import Buttons from './Buttons';
 import { useNavigate } from "react-router-dom";
-import { setAuthHeader, setRoleHeader } from '../helpers/axios_helper';
-import logo from '../img/logo.png';
+import { setAuthHeader, setRoleHeader } from '../../util/axios_helper';
+import logo from '../../img/logo.png';
 
 export default function Header() {
   const navigate = useNavigate();
