+++ gpb-front/src/components/game/GameContent.js
