+++ gpb-front/src/components/game/GameInfo.js
@@ -1,9 +1,12 @@
 import * as React from 'react'
-import Message from './Message';
+
+import '../../styles/gameInfo.css';
+
+import Message from '../../util/message';
 import { GameImage, GameAvailability } from './GameImage';
 import { useParams } from 'react-router-dom'
-import { request, isUserAdmin } from '../helpers/axios_helper';
-import { isUserAuth, defaultRequestErrorCheck } from '../helpers/axios_helper';
+import { isUserAdmin, isUserAuth } from '../../util/axios_helper';
+import { getGameRequest, subscribeForGameRequest, unsubscribeForGameRequest, removeGameRequest } from '../../request/gameRequests';
 import { useNavigate } from 'react-router-dom';
 
 export default function GameInfo(props) {
@@ -11,53 +14,10 @@ export default function GameInfo(props) {
     const [game, setGame] = React.useState(null);
     const navigate = useNavigate();
 
-    const handleError = (error) => {
-        defaultRequestErrorCheck(error)
-        if (error.response.status === 401) {
-            navigate(0);
-        }
-    };
-
     let { gameId } = useParams();
     React.useEffect(() => {
-        request('GET', '/game/' + gameId).then((response) => {
-            setGame(response.data);
-        }).catch(
-            (error) => {
-                handleError(error)
-            }
-        );
+        getGameRequest(gameId, setGame, navigate)
     }, []);
-
-    const subscribe = () => {
-        request('POST', '/user/games/' + gameId, {}).then((response) => {
-            navigate(0)
-        }).catch(
-            (error) => {
-                handleError(error)
-            }
-        );
-    }
-
-    const unsubscribe = () => {
-        request('DELETE', '/user/games/' + gameId, {}).then((response) => {
-            navigate(0)
-        }).catch(
-            (error) => {
-                handleError(error)
-            }
-        );
-    }
-
-    const removeGame = () => {
-        request('DELETE', '/game/' + gameId, {}).then((response) => {
-            navigate("/")
-        }).catch(
-            (error) => {
-                handleError(error)
-            }
-        );
-    }
 
     if (!game) return null;
 
@@ -70,14 +30,11 @@ export default function GameInfo(props) {
                     </div>
                     <div class="App-game-page-info-half">
                         <div class="App-game-page-info">
-
-
                             <div class="App-game-page-info-title">
                                 {game.name}
                             </div>
                             <div class="App-game-page-info-common  ">
                                 <div class="App-game-page-info-common-price">
-
                                     <GameAvailability available={game.available} />
                                     <div class="App-game-content-list-game-info-price">
                                         {game.minPrice} - {game.maxPrice} ₴
@@ -88,9 +45,8 @@ export default function GameInfo(props) {
                                     <GenreList genres={game.genres} />
                                 </div>
                             </div>
-                            <SubscribeButton isSubscribed={game.userSubscribed} subscribe={subscribe}
-                                unsubscribe={unsubscribe} />
-                                {isUserAdmin() && <RemoveButton removeGame={removeGame} /> }
+                            <SubscribeButton isSubscribed={game.userSubscribed} gameId={gameId} navigate={navigate} />
+                            {isUserAdmin() && <RemoveButton gameId={gameId} navigate={navigate} />}
                             <div class="App-game-page-info-storeList">
                                 <GameInStoreList stores={game.gamesInShop} />
                             </div>
@@ -99,10 +55,7 @@ export default function GameInfo(props) {
                 </div >
             </div >
         </div >
-
-
     );
-
 };
 
 function GenreList(props) {
@@ -111,14 +64,13 @@ function GenreList(props) {
         listItems.push(<div class="App-game-page-info-common-genre"><Message string={'app.game.genre.' + genre.toLowerCase()} /></div>)
     })
     return listItems
-
 }
 
 function GameInStoreList(props) {
     const listItems = [];
     props.stores.map(gameInStore => {
         let domain = (new URL(gameInStore.url)).hostname;
-        let image = require(`../img/${domain}.png`)
+        let image = require(`../../img/${domain}.png`)
         listItems.push(<a class="App-game-page-info-storeList-store " href={gameInStore.url} target="_blank">
             <img src={image} />
             <div class="">{domain}</div>
@@ -134,11 +86,10 @@ function GameInStoreList(props) {
 }
 
 function SubscribeButton(props) {
-
     return (
         <div class="App-game-page-info-subscribe">
-            <button type="submit" className="btn btn-primary btn-block mb-3"
-                disabled={!isUserAuth()} onClick={() => props.isSubscribed ? props.unsubscribe() : props.subscribe()}>
+            <button type="submit" className="btn btn-primary btn-block mb-3" disabled={!isUserAuth()}
+                onClick={() => props.isSubscribed ? unsubscribeForGameRequest(props.gameId, props.navigate) : subscribeForGameRequest(props.gameId, props.navigate)}>
                 {props.isSubscribed ? <Message string={'app.game.info.unsubscribe'} /> : <Message string={'app.game.info.subscribe'} />}
             </button>
             <span>{!isUserAuth() && <Message string={'app.game.info.need.auth'} />}</span>
@@ -147,10 +98,10 @@ function SubscribeButton(props) {
 }
 
 function RemoveButton(props) {
-
     return (
         <div class="App-game-page-info-subscribe">
-            <button type="submit" className="btn btn-primary btn-block mb-3 App-game-page-info-remove" onClick={props.removeGame}>
+            <button type="submit" className="btn btn-primary btn-block mb-3 App-game-page-info-remove"
+                onClick={removeGameRequest(props.gameId, props.navigate)}>
                 <Message string={'app.game.info.remove'} />
             </button>
         </div>
