+++ gpb-front/src/styles/gameInfo.css
@@ -1,0 +1,155 @@
+.App-game-page-template {
+    background-color: #13538881;
+    box-sizing: border-box;
+    width: 100%;
+    padding-top: 10px;
+    padding-bottom: 10px;
+    padding-right: 10px;
+    padding-left: 10px;
+    margin-right: auto;
+    margin-left: auto;
+    max-width: 1140px;
+    border-radius: 10px;
+    display: flex;
+  }
+  
+  .App-game-page {
+    width: 100%;
+    display: flex;
+    flex-direction: row;
+    justify-content: space-evenly;
+    align-items: center;
+  }
+  
+  .App-game-page-image {
+    flex: 0 0 50%;
+    max-width: 50%;
+  }
+  
+  .App-game-page-image img {
+    width: 85%;
+    height: auto;
+  }
+  
+  .App-game-page {
+    
+    display: flex;
+    flex-direction: row;
+    justify-content: space-evenly;
+    align-items: center;
+  }
+  
+  .App-game-page-info-half{
+    flex: 0 0 50%;
+    width: 50%;
+    height: 100%;
+  }
+  
+  .App-game-page-info {
+    height: 100%;
+    flex: 0 0 100%;
+    width: 100%;
+    display: flex;
+    flex-wrap: wrap;
+    align-content: flex-start;
+  }
+  
+  .App-game-page-info-title {
+    margin-top: 10px;
+    flex: 0 0 100%;
+    max-width: 100%;
+    font-size: 2.5rem;
+  }
+  
+  .App-game-page-info-common {
+    margin-top: 30px;
+    flex: 0 0 100%;
+    width: 100%;
+    display: flex;
+    flex-direction: row;
+    align-items: flex-start;
+  }
+  
+  .App-game-page-info-common-price {
+    flex: 0 0 50%;
+    max-width: 50%;
+  
+    display: flex;
+    flex-direction: column;
+    align-items: flex-start;
+  }
+  
+  .App-game-page-info-common-genre {
+    flex: 0 0 50%;
+    max-width: 50%;
+    display: flex;
+    flex-direction: column;
+    align-items: flex-start;
+  }
+  
+  .App-game-page-info-subscribe {
+    flex: 0 0 100%;
+    width: 100%;
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+  }
+  
+  .App-game-page-info-remove {
+    background-color: #ff0000 !important;
+    border-color: #ff0000 !important;
+  }
+  
+  .App-game-page-info-storeList {
+    margin-top: 30px;
+    flex: 0 0 100%;
+    width: 100%;
+    display: flex;
+    flex-direction: column;
+    justify-content: space-evenly;
+    align-items: center;
+  }
+  
+  .App-game-page-info-storeList-store {
+    max-height: 35px;
+    color: white;
+    text-decoration: none !important;
+    flex: 0 0 100%;
+    width: 100%;
+    border: 1px solid rgb(39, 47, 71);
+    display: flex;
+    flex-direction: row;
+    justify-content: space-between;
+    align-items: center;
+    padding-left: 10px;
+  }
+  
+  .App-game-page-info-storeList-store-price-section {
+    width: 25%;
+    display: flex;
+    flex-direction: row;
+    justify-content: space-around;
+    align-items: center;
+  }
+  
+  .App-game-page-info-storeList-store-price {
+    font-size: 0.9rem;
+    color: rgb(128, 126, 126);
+    text-decoration: line-through;
+  }
+  
+  .App-game-page-info-storeList-store-discount {
+    background-color: green;
+    height: 20px;
+    width: 40px;
+    display: flex;
+    align-items: center;
+  }
+  
+  .App-game-page-info-storeList-store-discountPrice {
+    font-size: 1.2rem;
+  }
+  
+  .App-game-page-info-storeList-store img {
+    max-width: 35px;
+  }
