+++ gpb-front/src/util/axios_helper.js
@@ -58,3 +58,4 @@ export const request = (method, url, data) => {
         data: data
     });
 };
+
