+++ gpb-front/src/util/message.js
